reorganized scaffolding under /src/store
