+++ packages/wallet-ui/package.json
@@ -22,8 +22,19 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "CI=true react-scripts test",
-    "test:coverage": "CI=true react-scripts test --coverage --collectCoverageFrom=src/**/*.ts* --collectCoverageFrom=!src/**/*.mock.ts --collectCoverageFrom=!src/**/*.d.ts --collectCoverageFrom=!src/**/types.ts --collectCoverageFrom=!src/**/index.ts --collectCoverageFrom=!src/**/call-action.ts --collectCoverageFrom=!src/**/log.ts",
+    "test:coverage": "CI=true react-scripts test --coverage",
     "eject": "react-scripts eject"
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "./src/**/*.ts*",
+      "!./src/**/*.mock.ts",
+      "!./src/**/*.d.ts",
+      "!./src/**/types.ts",
+      "!./src/**/index.ts",
+      "!./src/utils/log.ts",
+      "!./src/store/test-utils/*.ts"
+    ]
   },
   "eslintConfig": {
     "extends": "react-app"
