+++ packages/wallet-ui/src/store/store.ts
@@ -1,11 +1,9 @@
-import { combineReducers, createStore, applyMiddleware, compose } from "redux";
+import { applyMiddleware, combineReducers, compose, createStore } from "redux";
 import ReduxThunk from "redux-thunk";
-
-import { reducers as UserState } from "./user";
-import { reducers as WalletState } from "./wallet";
-import { reducers as ChannelsState } from "./channels";
-
+import { reducers as ChannelsState } from "./channels/channels";
 import { ApplicationState, StoreAction } from "./types";
+import { reducers as UserState } from "./user/user";
+import { reducers as WalletState } from "./wallet/wallet";
 
 export default createStore(
   combineReducers<ApplicationState, StoreAction<any, any>>({
