+++ packages/wallet-ui/src/store/user/user.mock.ts
@@ -3,14 +3,14 @@ import { parseEther } from "ethers/utils";
 import { History } from "history";
 import { Action } from "redux";
 import { ThunkAction } from "redux-thunk";
-import { RoutePath } from "../types";
+import { RoutePath } from "../../types";
 import {
   buildSignatureMessageForLogin,
   getUserFromStoredToken,
   storeTokenFromUser
-} from "../utils/counterfactual";
-import PlaygroundAPIClient from "../utils/hub-api-client";
-import { ActionType, ApplicationState, User } from "./types";
+} from "../../utils/counterfactual";
+import PlaygroundAPIClient from "../../utils/hub-api-client";
+import { ActionType, ApplicationState, User } from "../types";
 import { dispatchError, UserAddTransition } from "./user";
 
 export const USER_MOCK_DATA = {
