+++ packages/wallet-ui/src/store/user/user.ts
@@ -3,7 +3,7 @@ import { parseEther } from "ethers/utils";
 import { History } from "history";
 import { Action } from "redux";
 import { ThunkAction, ThunkDispatch } from "redux-thunk";
-import { RoutePath } from "../types";
+import { RoutePath } from "../../types";
 import {
   buildRegistrationSignaturePayload,
   buildSignatureMessageForLogin,
@@ -11,15 +11,15 @@ import {
   getNodeAddress,
   getUserFromStoredToken,
   storeTokenFromUser
-} from "../utils/counterfactual";
-import PlaygroundAPIClient, { ErrorDetail } from "../utils/hub-api-client";
+} from "../../utils/counterfactual";
+import PlaygroundAPIClient, { ErrorDetail } from "../../utils/hub-api-client";
 import {
   ActionType,
   ApplicationState,
   StoreAction,
   User,
   UserState
-} from "./types";
+} from "../types";
 
 const initialState = {
   user: {},
